move types package to cmdcommon
