+++ internal/cmdcommon/types/nullableBool.go
