+++ internal/kube/resources/resources.go
@@ -3,8 +3,8 @@ package resources
 import (
 	"context"
 
+	"github.com/kyma-project/cli.v3/internal/cmdcommon/types"
 	"github.com/kyma-project/cli.v3/internal/kube"
-	"github.com/kyma-project/cli.v3/internal/types"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
